etl/py3/requirements.txt
attrs==23.2.0
certifi==2024.6.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
dnspython==2.6.1
fiona==1.9.6
geopandas==0.14.4
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pymongo==4.7.3
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
shapely==2.0.4
six==1.16.0
tzdata==2024.1
